Refactor file name for better readability
