+++ librz/asm/arch/luac/v54/analysis_54.c
@@ -1,4 +1,4 @@
-#include "analysis_54.h"
+#include "arch_54.h"
 
 int lua54_anal_op(RzAnalysis *analysis, RzAnalysisOp *op, ut64 addr, const ut8 *data, int len) {
         if (!op) {
