--- librz/asm/arch/luac/v54/analysis_54.h
@@ -1,11 +1,0 @@
-#ifndef BUILD_ANALYSIS_54_H
-#define BUILD_ANALYSIS_54_H
-
-#include <rz_types.h>
-#include <rz_analysis.h>
-#include "librz/asm/arch/luac/luac_anal.h"
-#include "opcode_54.h"
-
-
-
-#endif //BUILD_ANALYSIS_54_H
