+++ librz/asm/meson.build
@@ -115,9 +115,9 @@ rz_asm_sources = [
   'arch/java/jvm.c',
   'arch/lanai/gnu/lanai-dis.c',
   'arch/lanai/gnu/lanai-opc.c',
-  'arch/luac/opcode.c',
+  'arch/luac/lua_arch.c',
   'arch/luac/v54/opcode_54.c',
-  'arch/luac/luac_dis.c',
+  'arch/luac/v54/disassembly_54.c',
   'arch/mips/gnu/mips-dis.c',
   'arch/mips/gnu/mips-opc.c',
   'arch/mips/gnu/mips16-opc.c',
