+++ librz/asm/p/asm_luac.c
@@ -1,11 +1,11 @@
-//
-// Created by heersin on 3/6/21.
-//
-
-#include "../arch/luac/luac_dis.h"
+#include "librz/asm/arch/luac/lua_arch.h"
 
 int rz_luac_disasm(RzAsm *a, RzAsmOp *opstruct, const ut8 *buf, int len) {
-	return luac_disasm(a, opstruct, buf, len);
+        LuaOpNameList oplist = get_lua54_opnames();
+        int r = lua54_disasm(opstruct, buf, len, oplist);
+        free_lua_opnames(oplist);
+        opstruct->size = r;
+        return r;
 }
 
 RzAsmPlugin rz_asm_plugin_luac = {
